+++ lib/ApiRequestor.php
@@ -84,7 +84,7 @@ class ApiRequestor
      *    hitting the API.
      * @throws Error\Api otherwise.
      */
-    public function handleApiError($rbody, $rcode, $rheaders, $resp)
+    public function handleErrorResponse($rbody, $rcode, $rheaders, $resp)
     {
         if (!is_array($resp) || !isset($resp['error'])) {
             $msg = "Invalid response object from API: $rbody "
@@ -92,33 +92,67 @@ class ApiRequestor
             throw new Error\Api($msg, $rcode, $rbody, $resp, $rheaders);
         }
 
-        $error = $resp['error'];
-        $msg = isset($error['message']) ? $error['message'] : null;
-        $param = isset($error['param']) ? $error['param'] : null;
-        $code = isset($error['code']) ? $error['code'] : null;
+        $errorData = $resp['error'];
+
+        $error = null;
+        if (is_string($errorData)) {
+            $error = self::_specificOAuthError($rbody, $rcode, $rheaders, $resp, $errorData);
+        }
+        if (!$error) {
+            $error = self::_specificAPIError($rbody, $rcode, $rheaders, $resp, $errorData);
+        }
+
+        throw $error;
+    }
+
+    private static function _specificAPIError($rbody, $rcode, $rheaders, $resp, $errorData)
+    {
+        $msg = isset($errorData['message']) ? $errorData['message'] : null;
+        $param = isset($errorData['param']) ? $errorData['param'] : null;
+        $code = isset($errorData['code']) ? $errorData['code'] : null;
 
         switch ($rcode) {
             case 400:
                 // 'rate_limit' code is deprecated, but left here for backwards compatibility
                 // for API versions earlier than 2015-09-08
                 if ($code == 'rate_limit') {
-                    throw new Error\RateLimit($msg, $param, $rcode, $rbody, $resp, $rheaders);
+                    return new Error\RateLimit($msg, $param, $rcode, $rbody, $resp, $rheaders);
                 }
 
                 // intentional fall-through
             case 404:
-                throw new Error\InvalidRequest($msg, $param, $rcode, $rbody, $resp, $rheaders);
+                return new Error\InvalidRequest($msg, $param, $rcode, $rbody, $resp, $rheaders);
             case 401:
-                throw new Error\Authentication($msg, $rcode, $rbody, $resp, $rheaders);
+                return new Error\Authentication($msg, $rcode, $rbody, $resp, $rheaders);
             case 402:
-                throw new Error\Card($msg, $param, $code, $rcode, $rbody, $resp, $rheaders);
+                return new Error\Card($msg, $param, $code, $rcode, $rbody, $resp, $rheaders);
             case 403:
-                throw new Error\Permission($msg, $rcode, $rbody, $resp, $rheaders);
+                return new Error\Permission($msg, $rcode, $rbody, $resp, $rheaders);
             case 429:
-                throw new Error\RateLimit($msg, $param, $rcode, $rbody, $resp, $rheaders);
+                return new Error\RateLimit($msg, $param, $rcode, $rbody, $resp, $rheaders);
             default:
-                throw new Error\Api($msg, $rcode, $rbody, $resp, $rheaders);
+                return new Error\Api($msg, $rcode, $rbody, $resp, $rheaders);
         }
+    }
+
+    private static function _specificOAuthError($rbody, $rcode, $rheaders, $resp, $errorCode)
+    {
+        $description = isset($resp['error_description']) ? $resp['error_description'] : $errorCode;
+
+        switch ($errorCode) {
+            case 'invalid_grant':
+                return new Error\OAuth\InvalidGrant($errorCode, $description, $rcode, $rbody, $resp, $rheaders);
+            case 'invalid_request':
+                return new Error\OAuth\InvalidRequest($errorCode, $description, $rcode, $rbody, $resp, $rheaders);
+            case 'invalid_scope':
+                return new Error\OAuth\InvalidScope($errorCode, $description, $rcode, $rbody, $resp, $rheaders);
+            case 'unsupported_grant_type':
+                return new Error\OAuth\UnsupportedGrantType($errorCode, $description, $rcode, $rbody, $resp, $rheaders);
+            case 'unsupported_response_type':
+                return new Error\OAuth\UnsupportedResponseType($errorCode, $description, $rcode, $rbody, $resp, $rheaders);
+        }
+
+        return null;
     }
 
     private static function _formatAppInfo($appInfo)
@@ -137,23 +171,12 @@ class ApiRequestor
         }
     }
 
-    private static function _defaultHeaders($apiKey)
+    private static function _defaultHeaders($apiKey, $clientInfo = null)
     {
-        $appInfo = Stripe::getAppInfo();
-
         $uaString = 'Stripe/v1 PhpBindings/' . Stripe::VERSION;
 
         $langVersion = phpversion();
         $uname = php_uname();
-
-        $httplib = 'unknown';
-        $ssllib = 'unknown';
-
-        if (function_exists('curl_version')) {
-            $curlVersion = curl_version();
-            $httplib = 'curl ' . $curlVersion['version'];
-            $ssllib = $curlVersion['ssl_version'];
-        }
 
         $appInfo = Stripe::getAppInfo();
         $ua = array(
@@ -162,9 +185,10 @@ class ApiRequestor
             'lang_version' => $langVersion,
             'publisher' => 'stripe',
             'uname' => $uname,
-            'httplib' => $httplib,
-            'ssllib' => $ssllib,
         );
+        if ($clientInfo) {
+            $ua = array_merge($clientInfo, $ua);
+        }
         if ($appInfo !== null) {
             $uaString .= ' ' . self::_formatAppInfo($appInfo);
             $ua['application'] = $appInfo;
@@ -193,9 +217,17 @@ class ApiRequestor
             throw new Error\Authentication($msg);
         }
 
+        // Clients can supply arbitrary additional keys to be included in the
+        // X-Stripe-Client-User-Agent header via the optional getUserAgentInfo()
+        // method
+        $clientUAInfo = null;
+        if (method_exists($this->httpClient(), 'getUserAgentInfo')) {
+            $clientUAInfo = $this->httpClient()->getUserAgentInfo();
+        }
+
         $absUrl = $this->_apiBase.$url;
         $params = self::_encodeObjects($params);
-        $defaultHeaders = $this->_defaultHeaders($myApiKey);
+        $defaultHeaders = $this->_defaultHeaders($myApiKey, $clientUAInfo);
         if (Stripe::$apiVersion) {
             $defaultHeaders['Stripe-Version'] = Stripe::$apiVersion;
         }
@@ -272,7 +304,7 @@ class ApiRequestor
         }
 
         if ($rcode < 200 || $rcode >= 300) {
-            $this->handleApiError($rbody, $rcode, $rheaders, $resp);
+            $this->handleErrorResponse($rbody, $rcode, $rheaders, $resp);
         }
         return $resp;
     }