+++ lib/HttpClient/CurlClient.php
@@ -36,6 +36,8 @@ class CurlClient implements ClientInterface
 
     protected $defaultOptions;
 
+    protected $userAgentInfo;
+
     /**
      * CurlClient constructor.
      *
@@ -52,11 +54,26 @@ class CurlClient implements ClientInterface
     public function __construct($defaultOptions = null)
     {
         $this->defaultOptions = $defaultOptions;
+        $this->initUserAgentInfo();
+    }
+
+    public function initUserAgentInfo()
+    {
+        $curlVersion = curl_version();
+        $this->userAgentInfo = array(
+            'httplib' =>  'curl ' . $curlVersion['version'],
+            'ssllib' => $curlVersion['ssl_version'],
+        );
     }
 
     public function getDefaultOptions()
     {
         return $this->defaultOptions;
+    }
+
+    public function getUserAgentInfo()
+    {
+        return $this->userAgentInfo;
     }
 
     // USER DEFINED TIMEOUTS
@@ -114,16 +131,16 @@ class CurlClient implements ClientInterface
             }
             $opts[CURLOPT_HTTPGET] = 1;
             if (count($params) > 0) {
-                $encoded = self::encode($params);
+                $encoded = Util\Util::urlEncode($params);
                 $absUrl = "$absUrl?$encoded";
             }
         } elseif ($method == 'post') {
             $opts[CURLOPT_POST] = 1;
-            $opts[CURLOPT_POSTFIELDS] = $hasFile ? $params : self::encode($params);
+            $opts[CURLOPT_POSTFIELDS] = $hasFile ? $params : Util\Util::urlEncode($params);
         } elseif ($method == 'delete') {
             $opts[CURLOPT_CUSTOMREQUEST] = 'DELETE';
             if (count($params) > 0) {
-                $encoded = self::encode($params);
+                $encoded = Util\Util::urlEncode($params);
                 $absUrl = "$absUrl?$encoded";
             }
         } else {
@@ -237,46 +254,5 @@ class CurlClient implements ClientInterface
     private static function caBundle()
     {
         return dirname(__FILE__) . '/../../data/ca-certificates.crt';
-    }
-
-    /**
-     * @param array $arr An map of param keys to values.
-     * @param string|null $prefix
-     *
-     * Only public for testability, should not be called outside of CurlClient
-     *
-     * @return string A querystring, essentially.
-     */
-    public static function encode($arr, $prefix = null)
-    {
-        if (!is_array($arr)) {
-            return $arr;
-        }
-
-        $r = array();
-        foreach ($arr as $k => $v) {
-            if (is_null($v)) {
-                continue;
-            }
-
-            if ($prefix) {
-                if ($k !== null && (!is_int($k) || is_array($v))) {
-                    $k = $prefix."[".$k."]";
-                } else {
-                    $k = $prefix."[]";
-                }
-            }
-
-            if (is_array($v)) {
-                $enc = self::encode($v, $k);
-                if ($enc) {
-                    $r[] = $enc;
-                }
-            } else {
-                $r[] = urlencode($k)."=".urlencode($v);
-            }
-        }
-
-        return implode("&", $r);
     }
 }