+++ tests/CurlClientTest.php
@@ -23,6 +23,15 @@ class CurlClientTest extends TestCase
         $this->assertSame(0, $curl->getConnectTimeout());
     }
 
+    public function testUserAgentInfo()
+    {
+        $curl = new CurlClient();
+        $uaInfo = $curl->getUserAgentInfo();
+        $this->assertNotNull($uaInfo);
+        $this->assertNotNull($uaInfo['httplib']);
+        $this->assertNotNull($uaInfo['ssllib']);
+    }
+
     public function testDefaultOptions()
     {
         // make sure options array loads/saves properly
@@ -46,48 +55,6 @@ class CurlClientTest extends TestCase
         });
         $this->setExpectedException('Stripe\Error\Api', "Non-array value returned by defaultOptions CurlClient callback");
         $withBadClosure->request('get', 'https://httpbin.org/status/200', array(), array(), false);
-    }
-
-    public function testEncode()
-    {
-        $a = array(
-            'my' => 'value',
-            'that' => array('your' => 'example'),
-            'bar' => 1,
-            'baz' => null
-        );
-
-        $enc = CurlClient::encode($a);
-        $this->assertSame('my=value&that%5Byour%5D=example&bar=1', $enc);
-
-        $a = array('that' => array('your' => 'example', 'foo' => null));
-        $enc = CurlClient::encode($a);
-        $this->assertSame('that%5Byour%5D=example', $enc);
-
-        $a = array('that' => 'example', 'foo' => array('bar', 'baz'));
-        $enc = CurlClient::encode($a);
-        $this->assertSame('that=example&foo%5B%5D=bar&foo%5B%5D=baz', $enc);
-
-        $a = array(
-            'my' => 'value',
-            'that' => array('your' => array('cheese', 'whiz', null)),
-            'bar' => 1,
-            'baz' => null
-        );
-
-        $enc = CurlClient::encode($a);
-        $expected = 'my=value&that%5Byour%5D%5B%5D=cheese'
-              . '&that%5Byour%5D%5B%5D=whiz&bar=1';
-        $this->assertSame($expected, $enc);
-
-        // Ignores an empty array
-        $enc = CurlClient::encode(array('foo' => array(), 'bar' => 'baz'));
-        $expected = 'bar=baz';
-        $this->assertSame($expected, $enc);
-
-        $a = array('foo' => array(array('bar' => 'baz'), array('bar' => 'bin')));
-        $enc = CurlClient::encode($a);
-        $this->assertSame('foo%5B0%5D%5Bbar%5D=baz&foo%5B1%5D%5Bbar%5D=bin', $enc);
     }
 
     public function testSslOption()